move the modal_shared folder outside of the redirect_notification directory
