+++ src/components/ExtensionMessage.js
@@ -1,11 +1,11 @@
-import Modal from "./redirect_notification/modal_shared/Modal";
-import Header from "./redirect_notification/modal_shared/Header";
-import Heading from "./redirect_notification/modal_shared/Heading";
-import Body from "./redirect_notification/modal_shared/Body";
-import MainImage from "./redirect_notification/modal_shared/MainImage";
+import Modal from "./modal_shared/Modal";
+import Header from "./modal_shared/Header";
+import Heading from "./modal_shared/Heading";
+import Body from "./modal_shared/Body";
+import MainImage from "./modal_shared/MainImage";
 import Footer from "../pages/info_page_shared/Footer";
-import ButtonContainer from "./redirect_notification/modal_shared/ButtonContainer";
-import GoToButton from "./redirect_notification/modal_shared/GoToButton";
+import ButtonContainer from "./modal_shared/ButtonContainer";
+import GoToButton from "./modal_shared/GoToButton";
 import { getExtensionInstallationLinks } from "../utils/misc/extensionCommunication";
 import FileDownloadOutlinedIcon from "@mui/icons-material/FileDownloadOutlined";
 import Feature from "../features/Feature";
