+++ src/components/modal_shared/Footer.js
