+++ src/components/redirect_notification/MobileNotification.js
@@ -1,13 +1,13 @@
-import * as s from "./modal_shared/Modal.sc";
+import * as s from "../modal_shared/Modal.sc";
 import { useState } from "react";
-import Modal from "./modal_shared/Modal";
-import Header from "./modal_shared/Header";
-import Heading from "./modal_shared/Heading";
-import Body from "./modal_shared/Body";
-import Footer from "./modal_shared/Footer";
-import ButtonContainer from "./modal_shared/ButtonContainer";
-import Checkbox from "./modal_shared/Checkbox";
-import GoToButton from "./modal_shared/GoToButton";
+import Modal from "../modal_shared/Modal";
+import Header from "../modal_shared/Header";
+import Heading from "../modal_shared/Heading";
+import Body from "../modal_shared/Body";
+import Footer from "../modal_shared/Footer";
+import ButtonContainer from "../modal_shared/ButtonContainer";
+import Checkbox from "../modal_shared/Checkbox";
+import GoToButton from "../modal_shared/GoToButton";
 import BookmarkBorderIcon from "@mui/icons-material/BookmarkBorder";
 
 export default function MobileNotification({
